chore: streamcreate refactor
